some changes

+++ emulation.h
@@ -2,7 +2,7 @@
 #define EMULATION_H
 
 //start emulation
-int emulate(short int* code, long int codesize);
+int emulate(int* code, long int codesize);
 
 //initialize emulator
 int initialize_emulator(long int size);
